requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
streamlit
pandas